feat: Refactor application structure by moving main logic to a dedicated App component
- Created a new `App.tsx` file to encapsulate the main application logic, including the document selector for resume and cover letter.
- Simplified `index.tsx` to serve as the entry point, importing the new `App` component.
- Retained the integration of Tailwind CSS and shadcn/ui components for consistent styling and UI functionality.
